Use BaseEntity

+++ App.Application/EntitiesCommandsQueries/Categories/Commands/CreateCategory/CreateCategoryCommand.cs
@@ -2,7 +2,7 @@
 
 namespace App.Application.EntitiesCommandsQueries.Categories.Commands.CreateCategory
 {
-    public class CreateCategoryCommand : IRequest<int>
+    public class CreateCategoryCommand : IRequest<long>
     {
         public string CategoryName { get; set; }
         public string Description { get; set; }
